requirements.txt
streamlit
pandas
numpy
matplotlib
missingno
scikit-learn
Pillow
requests
flask
tensorflow-hub
tensorflow
pickle